Use a double instead of an int as the sort value for method items
git-svn-id: https://smali.googlecode.com/svn/trunk@624 55b6fa8a-2a1e-11de-a435-ffa8d773f76a

+++ baksmali/src/main/java/org/jf/baksmali/Adaptors/BlankMethodItem.java
@@ -42,7 +42,7 @@ public class BlankMethodItem extends MethodItem {
         }
     }
 
-    public int getSortOrder() {
+    public double getSortOrder() {
         return Integer.MAX_VALUE;
     }
